Add Environment Layer to Full Mirror Quick Set Function (#23)
* Add Environment Layer to Full Mirror Quick Set

This change allows the "Show Players/World" functionality to also include `Environment`, in addition to `Default`, `Player`, and `MirrorReflection`. This was done because `Environment` is officially considered by VRChat as a Layer that can be used for Static Meshes in a World... as it shares the same exact Collision Matrix as `Default`.

In addition, I've added a new line to the Specific Layer Explanations detailing what `Environment` is used for. I've also done some proofreading as well.

* Update CustomMirrorEditor.cs